tests: Migrate AboutView tests to test_popups.
